+++ include/mp/flat/redef/std/range_con.h
@@ -245,7 +245,7 @@ protected:
     auto body = item.GetBody();
     body.add_term(1.0, slk);
     AlgConEQ lceq { std::move(body), item.ub() };
-    int i1 = GetMC().AddConstraint(std::move(lceq));
+    int i1 = (int)GetMC().AddConstraint(std::move(lceq));
     GetSlackLink().AddEntry({i, i1, slk});
   }
   void ConvertWithRhs(const ItemType& item, RangeRelations rr) {
